API Reference

Main Modules

SURD.surd — Method

surd(target::AbstractVector, agents::Tuple)

Calculate the Synergistic, Unique, and Redundant Decomposition (SURD) of information provided by multiple agents about a target variable.

Arguments

target: A vector representing the target variable.

agents: A tuple of vectors representing the agent variables.

log: (optional) If true, prints the SURD results to the console.

Utilities

SURD.Utils.particular_mutual_info — Method

particularmutualinfo(ptarget, pagents, pfull, targetindex, agent_indices)

Calculate the particular mutual information for a specific target bin and a set of agent indices.

Arguments

p_target: Histogram object for the target variable.

p_agents: Histogram object for the agent variables.

p_full: Histogram object for the full joint distribution of target and agents.

target_index: Index of the target bin for which to calculate the mutual information.

agent_indices: Indices of the agents to consider in the calculation.

SURD.Utils.prob_combination — Method

prob_combination(n::Int)

Generate all possible combinations of integers from 1 to n.

Arguments

n::Int: The upper limit of the range (inclusive).

SURD.Utils.shannon_entropy — Method

shannon_entropy(p::AbstractVector)

Calculate the Shannon entropy of a probability distribution represented by histogram p.

Arguments

p: Histogram object representing the probability distribution.